Frontend for Account Creation done
Modified AccountCreation/index.jsx to reflect recent design changes. A terms and conditions and privacy policy notice has been added. db_connect.php has been removed and will be worked on from scratch.
The frontend for account creation, on submit, will send a POST request to api/create-account with a json object in the following format:
{
  "firstName": "",
  "lastName": "",
  "gradMonth": "",
  "gradYear": "",
  "email": "",
  "promos": bool
}
The frontend does validation to enforce a @buffalo.edu email address and forces the user to aggree to the terms and conditions and privacy notice before submitting the form.
The frontend also does not let the user leave any fields blank (except the promos field which can be true/false).
On clicking submit, the user is met with a pop-up saying
"If an account with the email does not already exist, a temporary password has been sent to the email"
The pop-up gives the user an option to either stay on the same page, or go to log in.